Fixes for Debian Buster 10.5
nft in Buster was parsing the "-200" part of the add chain command
as an option, which resulted in the following error:

/usr/sbin/nft: invalid option -- '2'

To tell nft to not parser -200 as an option, we need to add the
double dash argument (--) right after the name of the nft binary
in the add chain command. (Got the solution from:
https://bugzilla.redhat.com/show_bug.cgi?id=1778883#c5)

Also, the "nft -j list sets <family>" command in Buster returns a
null object if there are no sets in the given family. So, just check
for a null object before going forward.

Should fix #1