Flush the filter sets only once before the update
The script was erroneously flushing the set every single time a
country's IP blocks was downloaded, which resulted in the filter
sets only having the IP blocks of the last country in the script's
argument list. This issue was brought up in #3 by @frankofno.

I have also modified the script to restore the old sets after a
flush in case we are unsuccessful in updating the filter sets with
the new IP blocks.