Reduce filter-chain's priority to -190
Since conntrack operations also use the priority -200, there was no
guarantee that conntrack operations would happen before traversing
the filter-chain rules (which was also using the priority -200).
Reducing filter-chain's priority to -190 fixes this issue.

Fixes a part of #5. Also need to change the script to add the ct state
rule as well.