Add the --allow-established flag
This flag adds a rule to allow packets from established connections. This
lets you create outgoing connections to IP addresses which are denied
by the filtering sets. Without this flag, packets from denied IP addresses
will always be dropped, whether the packet is part of an established
connection or not.

Fixes #5