Separate nft logging based on rule/policy action
Prior to this change, the `--log` could be used to add logging statements
to rules, in most cases this would log the dropped packets, when in
`--allow` mode however, it would log accepted packets, this made it hard
to configure what type of activity you wanted logged.

This change splits the `--log` into `--log-accept` & `--log-drop` to
allow the logging of accepted traffic, dropped or both if desired, to
support this the `--log-level` was split into `--log-accept-level` &
`--log-drop-level` and `--log-prefix` to `--log-accept-prefix` &
`--log-drop-prefix`.

To support this an extra final rule with an action matching the policy
will be added, if logging matching that policy ot counters are requested.

Co-authored-by: Ronnie P. Thomas <rpthms@users.noreply.github.com>